Split github-app into a folder module
Convert server/lib/github-app.ts (1192 lines) plus the two sibling
files github-app-webhook.ts and github-app-artifacts.ts into a
github-app/ folder with an index.ts barrel that re-exports the public
surface, so "../lib/github-app" import paths stay stable:

- config, oauth, jwt, api, http, persistence, validation
- webhook, webhook-gates (parsing, signature verify, gate DB ops)
- artifacts, artifacts-zip (egress policy, manifest, zip extraction)

Importers (routes/github-webhooks, workflow-gate-job,
release-candidate-pypi, tests) now resolve webhook/artifacts symbols
through the barrel. Crypto primitives import from ../crypto-utils.

Pure refactor, no behavior change.
