+++ server/lib/github-app/webhook.ts
@@ -1,15 +1,9 @@
-import { and, eq } from "drizzle-orm";
-import type { AppDb } from "../db";
-import { githubWorkflowGates } from "../db/schema";
-import {
-  fetchInstallationMetadata,
-  getInstallationAccessToken,
-  markInstallationStatus,
-  resolveDeploymentProtectionTarget,
-  type GithubAppConfig,
-  type InstallationRecord,
-  type ReleaseTargetRecord,
-} from "./github-app";
+import { hexDecode, hmacSha256, timingSafeEqual } from "../crypto-utils";
+import type { AppDb } from "../../db";
+import { getInstallationAccessToken } from "./api";
+import type { GithubAppConfig } from "./config";
+import { markInstallationStatus, resolveDeploymentProtectionTarget } from "./persistence";
+import { recordGateRequest, type WorkflowGateRecord } from "./webhook-gates";
 
 // ── Signature verification ───────────────────────────────────────────────────
 
@@ -204,235 +198,6 @@ function isObject(value: unknown): value is Record<string, unknown> {
   return Boolean(value && typeof value === "object" && !Array.isArray(value));
 }
 
-// ── Persistence ──────────────────────────────────────────────────────────────
-
-export type WorkflowGateStatus = "pending" | "approved" | "rejected" | "errored";
-
-export interface WorkflowGateRecord {
-  id: string;
-  organizationId: string;
-  installationRowId: string;
-  releaseTargetId: string;
-  deliveryId: string;
-  repositoryId: number;
-  repositoryFullName: string;
-  environment: string;
-  runId: number;
-  deploymentId: number | null;
-  deploymentCallbackUrl: string;
-  eventAction: string;
-  status: WorkflowGateStatus;
-  decision: "approved" | "rejected" | null;
-  decisionComment: string | null;
-  reportUrl: string | null;
-  scanId: string | null;
-  failureReason: string | null;
-  requestedAt: Date;
-  decidedAt: Date | null;
-  createdAt: Date;
-  updatedAt: Date;
-}
-
-interface RecordGateRequestInput {
-  deliveryId: string;
-  installation: InstallationRecord;
-  releaseTarget: ReleaseTargetRecord;
-  event: ParsedDeploymentProtectionEvent;
-}
-
-/**
- * Persist a pending gate after a `deployment_protection_rule.requested`
- * webhook. Returns the existing row if the same `X-GitHub-Delivery` ID has
- * already been recorded, so retries from GitHub are idempotent.
- */
-export async function recordGateRequest(
-  db: AppDb,
-  input: RecordGateRequestInput,
-): Promise<{ gate: WorkflowGateRecord; created: boolean }> {
-  const existing = await getGateByDeliveryId(db, input.deliveryId);
-  if (existing) return { gate: existing, created: false };
-
-  const now = new Date();
-  const id = crypto.randomUUID();
-  await db.insert(githubWorkflowGates).values({
-    id,
-    organizationId: input.installation.organizationId,
-    installationRowId: input.installation.id,
-    releaseTargetId: input.releaseTarget.id,
-    deliveryId: input.deliveryId,
-    repositoryId: input.event.repositoryId,
-    repositoryFullName: input.event.repositoryFullName,
-    environment: input.event.environment.toLowerCase(),
-    runId: input.event.runId,
-    deploymentId: input.event.deploymentId,
-    deploymentCallbackUrl: input.event.deploymentCallbackUrl,
-    eventAction: input.event.action,
-    status: "pending",
-    requestedAt: now,
-    createdAt: now,
-    updatedAt: now,
-  });
-  const fresh = await getGateByDeliveryId(db, input.deliveryId);
-  if (!fresh) throw new Error("workflow gate row vanished immediately after insert");
-  return { gate: fresh, created: true };
-}
-
-export async function getGateByDeliveryId(
-  db: AppDb,
-  deliveryId: string,
-): Promise<WorkflowGateRecord | null> {
-  const [row] = await db
-    .select()
-    .from(githubWorkflowGates)
-    .where(eq(githubWorkflowGates.deliveryId, deliveryId))
-    .limit(1);
-  return row ? readGateRow(row) : null;
-}
-
-export async function getGateForOrganization(
-  db: AppDb,
-  organizationId: string,
-  gateId: string,
-): Promise<WorkflowGateRecord | null> {
-  const [row] = await db
-    .select()
-    .from(githubWorkflowGates)
-    .where(
-      and(
-        eq(githubWorkflowGates.id, gateId),
-        eq(githubWorkflowGates.organizationId, organizationId),
-      ),
-    )
-    .limit(1);
-  return row ? readGateRow(row) : null;
-}
-
-export async function attachScanToGate(db: AppDb, gateId: string, scanId: string): Promise<void> {
-  const now = new Date();
-  await db
-    .update(githubWorkflowGates)
-    .set({ scanId, updatedAt: now })
-    .where(and(eq(githubWorkflowGates.id, gateId), eq(githubWorkflowGates.status, "pending")));
-}
-
-interface DecideGateInput {
-  gateId: string;
-  decision: "approved" | "rejected";
-  comment: string;
-  reportUrl?: string | null;
-}
-
-/**
- * Atomically transition a pending gate to `approved` or `rejected`. Returns
- * null if the gate was already decided (or never existed), so the caller can
- * skip re-posting to GitHub.
- */
-export async function markGateDecided(
-  db: AppDb,
-  input: DecideGateInput,
-): Promise<WorkflowGateRecord | null> {
-  const now = new Date();
-  const updated = await db
-    .update(githubWorkflowGates)
-    .set({
-      status: input.decision,
-      decision: input.decision,
-      decisionComment: input.comment,
-      reportUrl: input.reportUrl ?? null,
-      decidedAt: now,
-      updatedAt: now,
-    })
-    .where(and(eq(githubWorkflowGates.id, input.gateId), eq(githubWorkflowGates.status, "pending")))
-    .returning({ id: githubWorkflowGates.id });
-  if (updated.length === 0) return null;
-  const [row] = await db
-    .select()
-    .from(githubWorkflowGates)
-    .where(eq(githubWorkflowGates.id, input.gateId))
-    .limit(1);
-  return row ? readGateRow(row) : null;
-}
-
-export async function markGateErrored(
-  db: AppDb,
-  gateId: string,
-  reason: string,
-): Promise<WorkflowGateRecord | null> {
-  const now = new Date();
-  const updated = await db
-    .update(githubWorkflowGates)
-    .set({ failureReason: reason.slice(0, 500), updatedAt: now })
-    .where(and(eq(githubWorkflowGates.id, gateId), eq(githubWorkflowGates.status, "pending")))
-    .returning({ id: githubWorkflowGates.id });
-  if (updated.length === 0) return null;
-  const [row] = await db
-    .select()
-    .from(githubWorkflowGates)
-    .where(eq(githubWorkflowGates.id, gateId))
-    .limit(1);
-  return row ? readGateRow(row) : null;
-}
-
-function readGateRow(row: {
-  id: string;
-  organizationId: string;
-  installationRowId: string;
-  releaseTargetId: string;
-  deliveryId: string;
-  repositoryId: number;
-  repositoryFullName: string;
-  environment: string;
-  runId: number;
-  deploymentId: number | null;
-  deploymentCallbackUrl: string;
-  eventAction: string;
-  status: string;
-  decision: string | null;
-  decisionComment: string | null;
-  reportUrl: string | null;
-  scanId: string | null;
-  failureReason: string | null;
-  requestedAt: Date | string | number;
-  decidedAt: Date | string | number | null;
-  createdAt: Date | string | number;
-  updatedAt: Date | string | number;
-}): WorkflowGateRecord {
-  return {
-    id: row.id,
-    organizationId: row.organizationId,
-    installationRowId: row.installationRowId,
-    releaseTargetId: row.releaseTargetId,
-    deliveryId: row.deliveryId,
-    repositoryId: row.repositoryId,
-    repositoryFullName: row.repositoryFullName,
-    environment: row.environment,
-    runId: row.runId,
-    deploymentId: row.deploymentId,
-    deploymentCallbackUrl: row.deploymentCallbackUrl,
-    eventAction: row.eventAction,
-    status: normalizeGateStatus(row.status),
-    decision: normalizeGateDecision(row.decision),
-    decisionComment: row.decisionComment,
-    reportUrl: row.reportUrl,
-    scanId: row.scanId,
-    failureReason: row.failureReason,
-    requestedAt: new Date(row.requestedAt),
-    decidedAt: row.decidedAt ? new Date(row.decidedAt) : null,
-    createdAt: new Date(row.createdAt),
-    updatedAt: new Date(row.updatedAt),
-  };
-}
-
-function normalizeGateStatus(value: string): WorkflowGateStatus {
-  if (value === "approved" || value === "rejected" || value === "errored") return value;
-  return "pending";
-}
-
-function normalizeGateDecision(value: string | null): "approved" | "rejected" | null {
-  if (value === "approved" || value === "rejected") return value;
-  return null;
-}
-
 // ── Webhook dispatcher ───────────────────────────────────────────────────────
 
 export type WebhookOutcome =
@@ -542,38 +307,4 @@ export async function postDeploymentProtectionDecision(
       `GitHub deployment protection decision failed (${response.status}): ${text.slice(0, 200)}`,
     );
   }
-}
-
-// Re-export for convenience in callers that want a one-shot freshness check.
-export { fetchInstallationMetadata };
-
-// ── Crypto helpers ───────────────────────────────────────────────────────────
-
-async function hmacSha256(secret: string, message: string): Promise<Uint8Array> {
-  const key = await crypto.subtle.importKey(
-    "raw",
-    new TextEncoder().encode(secret),
-    { name: "HMAC", hash: "SHA-256" },
-    false,
-    ["sign"],
-  );
-  const signature = await crypto.subtle.sign("HMAC", key, new TextEncoder().encode(message));
-  return new Uint8Array(signature);
-}
-
-function timingSafeEqual(a: Uint8Array, b: Uint8Array): boolean {
-  if (a.length !== b.length) return false;
-  let diff = 0;
-  for (let i = 0; i < a.length; i += 1) diff |= a[i] ^ b[i];
-  return diff === 0;
-}
-
-function hexDecode(value: string): Uint8Array | null {
-  if (value.length === 0 || value.length % 2 !== 0) return null;
-  if (!/^[0-9a-fA-F]+$/.test(value)) return null;
-  const out = new Uint8Array(value.length / 2);
-  for (let i = 0; i < out.length; i += 1) {
-    out[i] = Number.parseInt(value.slice(i * 2, i * 2 + 2), 16);
-  }
-  return out;
 }
