+++ test/workers/github-webhook.test.ts
@@ -4,13 +4,14 @@ import { afterEach, describe, expect, test, vi } from "vitest";
 import { createDb, ensurePersonalOrganization } from "../../server/db";
 import * as schema from "../../server/db/schema";
 import { eq } from "drizzle-orm";
-import { createReleaseTarget, upsertInstallation } from "../../server/lib/github-app";
 import {
+  createReleaseTarget,
   getGateByDeliveryId,
   getGateForOrganization,
   markGateDecided,
   markGateErrored,
-} from "../../server/lib/github-app-webhook";
+  upsertInstallation,
+} from "../../server/lib/github-app";
 import { githubWebhookRoutes } from "../../server/routes/github-webhooks";
 import type { Bindings, Variables } from "../../server/types";
 
@@ -464,8 +465,7 @@ describe("markGateDecided", () => {
 
 describe("postDeploymentProtectionDecision integration via mock fetch", () => {
   test("rejects when the stored callback URL is no longer github-controlled", async () => {
-    const { postDeploymentProtectionDecision } =
-      await import("../../server/lib/github-app-webhook");
+    const { postDeploymentProtectionDecision } = await import("../../server/lib/github-app");
     const { readGithubAppConfig } = await import("../../server/lib/github-app");
     const config = readGithubAppConfig({
       GITHUB_APP_ID: "12345",
@@ -491,8 +491,7 @@ describe("postDeploymentProtectionDecision integration via mock fetch", () => {
   });
 
   test("posts to the callback URL with installation token + structured body", async () => {
-    const { postDeploymentProtectionDecision } =
-      await import("../../server/lib/github-app-webhook");
+    const { postDeploymentProtectionDecision } = await import("../../server/lib/github-app");
     const tokenFetch = vi.fn(async () =>
       Response.json({ token: "ghs_install_token", expires_at: "2099-01-01T00:00:00Z" }),
     );
